requirements.txt
mysql-connector==2.1.6
pyserial==3.4
SQLAlchemy==1.2.12
